Faster vectors (#16)
* Pretty much accelerated linear algebra operations.
By delegating to primitive Java.

* Pretty much accelerated linear algebra operations.
By delegating to primitive Java.

* One import warning fixed

* Added test tool, added a butterfly 3d sample.

* Sammon - fixed zero denominator issue.
It still diverges.

Butterfly moded to TestTools. Added a sample with 3d-cube.

* Another compiler error - "no position"
Will report

* Sammon kind of works, at least on the butterfly example.

Tuned up Sammon. Not sure if it will work as good on other examples.
Moved visualization to Viz object in ml/dimreduction.

* Sammon dimension reduction.

Added a test case for exact cube; must be solvable.

* Sammon - solved the problem with two projections mapping to the same point.

Since in this case the gradient is zero, we have to twist the point a little bit.
Twist it in a direction of unit cube vertices around the point.

* Sammon - solved the problem with two projections mapping to the same point.

Since in this case the gradient is zero, we have to twist the point a little bit.
Twist it in a direction of unit cube vertices around the point.

* Sammon - memory footprint improvement (somehow makes it faster, too)

Also, parallelized. Also, fixed a couple of bugs in parallelization.

* WIP: ml - adding GradientDescentEngine

What we currently have is search along gradient line

* WIP: ml - GradientDescentEngine, tested "search along gradient"

next: switch gradient.

* WIP: ml - GradientDescentEngine, a bunch of tests passes.

TODO: clean up.

* WIP: ml - GradientDescentEngine, a bunch of tests passes.

TODO: clean up.

* ml - GradientDescentEngine, done with it.

* ml - GradientDescentEngine, renamed, error => value

* ml - GradientDescentEngine, restructuring.

Renamed type parameters. State is not a special case of Gradient (physicists know).
Etc. Just make sure that State is mutable.

* ml - renamed Gradient to Tangent type;

Experimented with immutability. Sucks. Like 10x slower.
A little bit of type improvement in SKipGramModelTest.

* ml - Sammon dimension reducer now uses the new gradient search.